Refactor: Move secret models to a dedicated model package
Centralized `ApiKeyEntryData` and `Entry` types in the new `model` package to improve code organization and separation of concerns. Updated all references across services, providers, and handlers to reflect the new import path.
