+++ security/secret/provider.go
@@ -3,21 +3,8 @@ package secret
 import (
 	goservectx "github.com/softwareplace/goserve/context"
 	goservejwt "github.com/softwareplace/goserve/security/jwt"
-	"time"
+	"github.com/softwareplace/goserve/security/model"
 )
-
-type ApiKeyEntryData struct {
-	ClientName string        `json:"clientName"` // Client information for which the public key is generated (required)
-	Expiration time.Duration `json:"expiration"` // Expiration specifies the duration until the API key expires (optional).
-	ClientId   string        `json:"clientId"`   // ClientId represents the unique identifier for the client associated with the API key entry (required).
-}
-
-type Entry struct {
-	Key        string
-	Expiration time.Duration
-	Roles      []string
-	PublicKey  *string
-}
 
 // Provider is an interface designed to provide secure access to API secret keys
 // based on the context of an incoming API request.
@@ -58,7 +45,7 @@ type Provider[T goservectx.Principal] interface {
 	// Returns:
 	//   - jwt.Entry: The generated ApiJWTInfo object containing JWT details necessary for creating the API secret JWT.
 	//   - error: If an error occurs during the process, it returns the corresponding error; otherwise, nil.
-	GetJwtEntry(apiKeyEntryData ApiKeyEntryData, ctx *goservectx.Request[T]) (Entry, error)
+	GetJwtEntry(apiKeyEntryData model.ApiKeyEntryData, ctx *goservectx.Request[T]) (model.Entry, error)
 
 	// OnGenerated is invoked after an API key has been successfully generated.
 	// This function allows additional processing or handling, such as logging,
@@ -69,7 +56,7 @@ type Provider[T goservectx.Principal] interface {
 	//   - jwtEntry: The requested jwt.Entry.
 	//   - ctx: The API request context, containing metadata and principal
 	//		  information related to the API key generation.
-	OnGenerated(response goservejwt.Response, jwtEntry Entry, ctx goservectx.SampleContext[T])
+	OnGenerated(response goservejwt.Response, jwtEntry model.Entry, ctx goservectx.SampleContext[T])
 
 	// RequiredScopes returns a list of scopes that are mandatory for accessing the related end point
 	// resources registered by invoking server.Api SecretService method.
