+++ server/server_test.go
@@ -1,0 +1,9 @@
+package server
+
+import "testing"
+
+func TestServerImplementation(t *testing.T) {
+	t.Run("should add resource path api-key/generate when secret service was added and apiSecretKeyGeneratorResourceEnable is true", func(t *testing.T) {
+
+	})
+}
